requirements.txt
discord.py==2.6.4
aiohttp>=3.9.0
SQLAlchemy>=2.0.23
aiosqlite>=0.19.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
mcp>=0.5.0
cryptography>=41.0.0
cachetools>=5.3.0
nicegui>=2.0.0
google-generativeai>=0.8.0
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
duckduckgo-search>=8.0.0
tavily-python>=0.3.0
psutil>=5.9.0